This fixes NCL-811. A temp variable was being malloced to wrong size.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11868 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/nfp/relhumW.c
@@ -159,7 +159,7 @@ NhlErrorTypes relhum_W( void )
  */
   if(type_rh == NCL_float) {
     rh     = (void*)calloc(total_size_rh,sizeof(float));
-    tmp_rh = (void*)calloc(1,sizeof(float));
+    tmp_rh = (void*)calloc(1,sizeof(double));
     if(tmp_rh == NULL || rh == NULL) {
       NhlPError(NhlFATAL,NhlEUNKNOWN,"relhum: Unable to allocate memory for output array");
       return(NhlFATAL);
